# clients/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "node-flow",
    "version": "1",
    "description": "Workflow orchestration. Definitions are declarative JSON; workers are language-agnostic and lease tasks from a queue."
  },
  "components": {
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Long-lived, revocable, scoped. For CLIs and CI."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived token from POST /v1/auth/token. For worker fleets: a captured token is useless within the hour."
      }
    }
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/v1/ns/{ns}/queues": {
      "get": {
        "operationId": "queue_overview",
        "summary": "All queues currently holding work, with depth and lease holders",
        "tags": [
          "workers"
        ],
        "x-required-scopes": [
          "executions:read"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/ns/{ns}/queues/{queue}/depth": {
      "get": {
        "operationId": "queue_depth",
        "summary": "Queue depth — the real signal for autoscaling a worker fleet",
        "tags": [
          "workers"
        ],
        "x-required-scopes": [
          "executions:read"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "queue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/ns/{ns}/queues/{queue}/lease": {
      "post": {
        "operationId": "queue_lease",
        "summary": "Lease a batch of tasks, optionally waiting for work to appear",
        "description": "Requires the queue-specific scope `queues:lease:<queue>`. Admission control — concurrency caps, rate limits, semaphores — is applied here, so a worker cannot obtain more than its share by asking differently. Set `waitSeconds` to hold the request open until a task arrives, which removes the trade-off between poll frequency and task-start latency.",
        "tags": [
          "workers"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "queue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workerId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "count": {
                    "default": 1,
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100
                  },
                  "leaseSeconds": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 3600
                  },
                  "waitSeconds": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 3600
                  }
                },
                "required": [
                  "workerId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/ns/{ns}/queues/workers": {
      "get": {
        "operationId": "queue_workers",
        "summary": "Workers seen polling in the last 24 hours, by queue",
        "tags": [
          "workers"
        ],
        "x-required-scopes": [
          "executions:read"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
...
    "/v1/ns/{ns}/tasks/{taskId}/heartbeat": {
      "post": {
        "operationId": "task_heartbeat",
        "summary": "Extend the lease on a long-running task",
        "description": "409 if the lease is gone, so a worker can abandon work that would be refused.",
        "tags": [
          "workers"
        ],
        "x-required-scopes": [
          "tasks:report"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "queueName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "leaseToken": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "leaseSeconds": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 3600
                  }
                },
                "required": [
                  "queueName",
                  "leaseToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/ns/{ns}/tasks/{taskId}/logs": {
      "post": {
        "operationId": "task_appendLogs",
        "summary": "Append log lines to a running task",
        "description": "Only the worker holding the lease may write. Lines are capped in length and per task; beyond the cap they are dropped with a marker line saying so.",
        "tags": [
          "workers"
        ],
        "x-required-scopes": [
          "tasks:report"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workflowId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "leaseToken": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "logs": {
                    "minItems": 1,
                    "maxItems": 500,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 64000
                        },
                        "level": {
                          "type": "string",
                          "enum": [
                            "debug",
                            "info",
                            "warn",
                            "error"
                          ]
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  }
                },
                "required": [
                  "workflowId",
                  "leaseToken",
                  "logs"
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/ns/{ns}/tasks/{taskId}/report": {
      "post": {
        "operationId": "task_report",
        "summary": "Report a terminal result",
        "description": "Fenced on the lease token: a worker whose lease expired cannot overwrite the result of whoever holds the task now.",
        "tags": [
          "workers"
        ],
        "x-required-scopes": [
          "tasks:report"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "queueName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "workflowId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "leaseToken": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "COMPLETED",
                      "FAILED",
                      "FAILED_WITH_TERMINAL_ERROR"
                    ]
                  },
                  "output": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {}
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 4000
                  }
                },
                "required": [
                  "queueName",
                  "workflowId",
                  "leaseToken",
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
...
  }
}